Make compatible with PostgreSQL apt repositories
Sometimes the server installs postgresql from the postgresql apt
repositories rather than the Debian repositories (this is not currently
done in this role but elsewhere). In this case we shouldn't be
installing the "postgresql" apt package, because it has a dependency on
the latest version of postgresql and may change.